refactor + add test to reedsolomon encoding
