+++ src/frida_data.rs
@@ -1,6 +1,6 @@
 use crate::frida_error::FridaError;
 use core::mem;
-use winter_math::{fft, get_power_series, polynom, FieldElement, StarkField};
+use winter_math::{fft, polynom, FieldElement, StarkField};
 
 pub fn encoded_data_element_count<E: FieldElement>(data_size: usize) -> usize {
     let element_size = E::ELEMENT_BYTES - 1;
@@ -61,36 +61,36 @@ pub fn build_evaluations_from_data<E: FieldElement>(
     let mut symbols: Vec<E> = data_to_field_element(&encoded_data, domain_size)?;
     symbols.resize(domain_size / blowup_factor, E::default());
 
-    let inv_twiddles = fft::get_inv_twiddles(symbols.len());
-    fft::interpolate_poly(&mut symbols, &inv_twiddles);
+    // let inv_twiddles = fft::get_inv_twiddles(symbols.len());
+    // fft::interpolate_poly(&mut symbols, &inv_twiddles);
 
-    symbols.resize(domain_size, E::default());
-    let twiddles = fft::get_twiddles(domain_size);
-    fft::evaluate_poly(&mut symbols, &twiddles);
+    // symbols.resize(domain_size, E::default());
+    // let twiddles = fft::get_twiddles(domain_size);
+    // fft::evaluate_poly(&mut symbols, &twiddles);
 
-    Ok(symbols)
+    // Ok(symbols)
+    Ok(reed_solomon_encode_data(
+        &symbols,
+        domain_size / blowup_factor,
+        blowup_factor,
+    ))
 }
 
-pub fn build_evaluations_from_data2<E: FieldElement>(
-    raw_data: &[u8],
-    domain_size: usize,
+pub fn reed_solomon_encode_data<E: FieldElement>(
+    data: &[E],
+    ori_domain_size: usize,
     blowup_factor: usize,
-) -> Result<Vec<E>, FridaError> {
-    let ori_domain_size = domain_size / blowup_factor;
-    if !ori_domain_size.is_power_of_two() {
-        return Err(FridaError::BadDataLength());
-    }
-
-    let encoded_data = encode_data::<E>(raw_data, domain_size, blowup_factor);
-    let mut symbols: Vec<E> = data_to_field_element(&encoded_data, ori_domain_size)?;
-
+) -> Vec<E> {
     let inv_twiddles = fft::get_inv_twiddles(ori_domain_size);
-    fft::interpolate_poly(&mut symbols, &inv_twiddles); // symbols are now coeffecients of the polynomial
+    let mut coeffs = data.to_vec();
+    fft::interpolate_poly(&mut coeffs, &inv_twiddles);
 
-    let g_after_rs = E::from(E::BaseField::get_root_of_unity(domain_size.ilog2()));
-    let positions_after_rs: Vec<E> = get_power_series(g_after_rs, domain_size);
-    let final_evaluation = polynom::eval_many(&symbols, &positions_after_rs);
-    Ok(final_evaluation)
+    let domain_size = ori_domain_size * blowup_factor;
+    coeffs.resize(domain_size, E::default());
+    let twiddles = fft::get_twiddles(domain_size);
+    fft::evaluate_poly(&mut coeffs, &twiddles);
+
+    coeffs // coeffs is not the evaluation points
 }
 
 fn reconstruct_evaluations<E: FieldElement>(
@@ -183,7 +183,7 @@ pub fn recover_data_from_evaluations<E: FieldElement>(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use winter_math::{fields::f128::BaseElement, get_power_series};
+    use winter_math::fields::f128::BaseElement;
 
     #[test]
     fn test_build_evaluations_from_data() {
@@ -245,37 +245,24 @@ mod tests {
 
     #[test]
     fn test_reed_solomon_encoding() {
-        // create evaluation of N size where N is 2-adicity for roots of unity, else pad with zeroes
-        // blowup factor
-
-        // interpolate evaluation
-        // evaluate polynomial to see if result is the same
-        // positions = (0..N).map(|i| g^i)
-        let blowup_factor = 8;
-
-        let mut example_evaluation = vec![
+        let example_evaluation = vec![
             BaseElement::new(1),
             BaseElement::new(2),
             BaseElement::new(3),
+            BaseElement::new(4),
         ];
-        let next_power_of_two = example_evaluation.len().next_power_of_two();
-        while example_evaluation.len() < next_power_of_two {
-            example_evaluation.push(BaseElement::ZERO);
-        }
 
-        let g_before_rs = BaseElement::get_root_of_unity(next_power_of_two.ilog2());
+        let blowup_factor = 8;
+        let reed_solomon_encoded_evaluation =
+            reed_solomon_encode_data(&example_evaluation, example_evaluation.len(), blowup_factor);
 
-        let inv_twiddles = fft::get_inv_twiddles::<BaseElement>(next_power_of_two);
-        let mut coeffs: Vec<BaseElement> = example_evaluation.clone();
-        let positions_before_rs = get_power_series(g_before_rs, next_power_of_two);
-        fft::interpolate_poly(&mut coeffs, &inv_twiddles);
-        let before_evalution = polynom::eval_many(&coeffs, &positions_before_rs);
+        let positions = (0..4).map(|i| i * blowup_factor).collect::<Vec<_>>();
+        let rs_evaluations = positions
+            .iter()
+            .map(|p| reed_solomon_encoded_evaluation[*p])
+            .collect::<Vec<_>>();
 
-        let domain_size = next_power_of_two * blowup_factor;
-        let g_after_rs = BaseElement::get_root_of_unity(domain_size.ilog2());
-        let positions_after_rs = get_power_series(g_after_rs, domain_size);
-        let final_evaluation = polynom::eval_many(&coeffs, &positions_after_rs);
-        println!("final_evaluation: {:?}", final_evaluation);
+        assert_eq!(example_evaluation, rs_evaluations);
     }
 
     #[test]
