+++ src/frida_verifier/test.rs
@@ -2,7 +2,7 @@
 mod test {
     use crate::frida_prover_channel::{BaseProverChannel, BaseProverChannelTest};
     use crate::frida_random::{FridaRandom, FridaRandomCoin};
-    use crate::frida_verifier::FridaVerifier;
+    use crate::frida_verifier::verifier_deprecated::FridaVerifierDeprecated;
     use crate::utils::{build_evaluations, build_prover_channel};
     use winter_crypto::hashers::Blake3_256;
     use winter_fri::{DefaultVerifierChannel, FriOptions, FriProver};
@@ -46,7 +46,8 @@ mod test {
         let mut coin = FridaRandom::<Blake3, Blake3, BaseElement>::new(&[123]);
 
         let verifier =
-            FridaVerifier::new(&mut channel, &mut coin, options.clone(), max_degree).unwrap();
+            FridaVerifierDeprecated::new(&mut channel, &mut coin, options.clone(), max_degree)
+                .unwrap();
 
         let layer_alpha = verifier.layer_alphas();
